refactor(pano): split Pano.ts into posts/comments planes + stats (#34)
`Pano.ts` was 1,919 lines holding both planes of the service in one
`PanoLive` layer. Split the service-internal closures along the natural
posts-vs-comments seam into cohesive modules, leaving the `Pano` tag and its
public surface byte-identical to callers — a pure behavior-preserving refactor
(every method body + `Effect.fn` span name moved verbatim, no signature or wire
change).

New modules under `features/pano/`:
- `post-operations.ts` — posts plane: `makePostOperations(deps)` factory + the
  post Input/Result types, constants, and submit-validation gates.
- `comment-operations.ts` — comments plane: `makeCommentOperations(deps)` +
  the comment types, `SILINDI_PLACEHOLDER`, and the `[silindi]` tombstone.
- `pano-stats.ts` — the cross-plane `pano_stats` fold + `makePersistPanoStats`
  port (spans both planes, so it has its own home).
- `excerpt.ts` — the shared tweet-sized excerpt both planes derive.

`Pano.ts` keeps the `Pano` class unchanged and `PanoLive` becomes the thin
wiring seam: build the shared deps once, hand them to each plane's factory,
spread the two closure sets into the one service object. All prior `Pano.ts`
exports are re-exported so resolvers/tests/cross-feature consumers are
untouched. Sandbox-visibility (#1205) guards preserved verbatim.

typecheck + lint:worktree green; full unit tier (619 tests, incl. all pano
unit suites + domain-error-boundary sweep) green.

Closes #34
